+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 ### Updated
 
+- don't return trailing newline after comment in get statementish
 - `dump()` uses color when printing to a terminal and formats a little better
 - improve indent infer to check solo block statements, `ExceptHandler`s and `match_case`s
 - change `ExceptHandler`s and `match_case`s to use explicit slice type instead of `Module`